# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pharmacy POS API",
    "version": "1.0.0",
    "description": "Generated from shared Zod schemas (SSOT)"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
...
  "paths": {
...
    "/customers": {
      "get": {
        "summary": "List customers",
        "operationId": "listCustomers",
        "tags": [
          "Customers"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "examples": {
                  "customers": {
                    "summary": "Customer list",
                    "value": {
                      "success": true,
                      "message": "OK",
                      "data": [
                        {
                          "_id": "64f0a1b2c3d4e5f678901234",
                          "code": "C00001",
                          "name": "John Doe",
                          "membershipLevel": "regular"
                        }
                      ],
                      "timestamp": "2025-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a customer",
        "operationId": "createCustomer",
        "tags": [
          "Customers"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "examples": {
                  "created": {
                    "summary": "Created",
                    "value": {
                      "success": true,
                      "message": "Created",
                      "data": {
                        "_id": "64f0a1b2c3d4e5f678901234",
                        "code": "C00001",
                        "name": "John Doe"
                      },
                      "timestamp": "2025-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/customers/{id}": {
      "get": {
        "summary": "Get customer by ID",
        "operationId": "getCustomerById",
        "tags": [
          "Customers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "examples": {
                  "customer": {
                    "summary": "Customer",
                    "value": {
                      "success": true,
                      "message": "OK",
                      "data": {
                        "_id": "64f0a1b2c3d4e5f678901234",
                        "code": "C00001",
                        "name": "John Doe"
                      },
                      "timestamp": "2025-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a customer",
        "operationId": "updateCustomer",
        "tags": [
          "Customers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "examples": {
                  "updated": {
                    "summary": "Updated",
                    "value": {
                      "success": true,
                      "message": "Updated",
                      "data": {
                        "_id": "64f0a1b2c3d4e5f678901234",
                        "code": "C00001",
                        "name": "John Doe",
                        "membershipLevel": "gold"
                      },
                      "timestamp": "2025-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a customer",
        "operationId": "deleteCustomer",
        "tags": [
          "Customers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "examples": {
                  "deleted": {
                    "summary": "Deleted",
                    "value": {
                      "success": true,
                      "message": "Deleted",
                      "data": null,
                      "timestamp": "2025-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {},
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string"
                },
                "param": {
                  "type": "string"
                },
                "location": {
                  "type": "string"
                }
              },
              "required": [
                "msg"
              ]
            }
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          },
          "statusCode": {
            "type": "number"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
      "CustomerCreateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "code": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "pattern": "^[\\d\\s\\-+()]+$",
            "minLength": 8,
            "maxLength": 20
          },
          "email": {
            "type": "string",
            "pattern": "^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$",
            "maxLength": 254
          },
          "address": {
            "type": "string"
          },
          "idCardNumber": {
            "type": "string",
            "pattern": "^[A-Z][12]\\d{8}$"
          },
          "birthdate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "string",
                "format": "date-time"
              }
            ]
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other"
            ]
          },
          "allergies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "membershipLevel": {
            "type": "string",
            "enum": [
              "regular",
              "silver",
              "gold",
              "platinum"
            ]
          },
          "medicalHistory": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "additionalProperties": false
      },
      "CustomerUpdateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "code": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "pattern": "^[\\d\\s\\-+()]+$",
            "minLength": 8,
            "maxLength": 20
          },
          "email": {
            "type": "string",
            "pattern": "^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$",
            "maxLength": 254
          },
          "address": {
            "type": "string"
          },
          "idCardNumber": {
            "type": "string",
            "pattern": "^[A-Z][12]\\d{8}$"
          },
          "birthdate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "string",
                "format": "date-time"
              }
            ]
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other"
            ]
          },
          "allergies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "membershipLevel": {
            "type": "string",
            "enum": [
              "regular",
              "silver",
              "gold",
              "platinum"
            ]
          },
          "medicalHistory": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}